Applied HARMONY-3329 [drlvm][winx64][jit] Jitrino WIN64 enabling fixes Tested on SLES10@ia32, SLES9@x64, WinXP@ia32, IPF build.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@517206 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/codegenerator/ia32/Ia32CodeEmitter.cpp
@@ -530,9 +530,9 @@ void CodeEmitter::postPass()
 
                     uint8* blockStartIp = (uint8*)bb->getCodeStartAddr();
                     uint8* ip = movInst->emit(instCodeStartAddr);
-                    movInst->setCodeOffset(instCodeStartAddr - blockStartIp);
+                    movInst->setCodeOffset((uint32)(instCodeStartAddr - blockStartIp));
                     inst->emit(ip);
-                    inst->setCodeOffset(ip - blockStartIp);
+                    inst->setCodeOffset((uint32)(ip - blockStartIp));
 
                     if (bcOffset != ILLEGAL_VALUE) {
                         bcMap->setEntry((uint64)(POINTER_SIZE_INT)instCodeStartAddr, bcOffset); // MOV
